update close button and start shuffle funk
