+++ client/src/components/customPrompts/CustomPrompts.module.scss
@@ -1,11 +1,3 @@
 .wrapper {
-}
-
-.promptInput {
   display: flex;
-}
-
-.listedPrompt {
-  display: flex;
-  align-items: center;
 }
