--- client/src/components/prompt/prompt.scss
@@ -1,119 +1,0 @@
-@import '../../styles/mixins.scss';
-@import '../../styles/variables.scss';
-@import '../../styles/base.scss';
-
-.prompt {
-	display: flex;
-	align-items: center;
-	flex-direction: column;
-	font-family: 'Poppins', sans-serif;
-	font-style: italic;
-	justify-content: space-between;
-	text-align: center;
-	font-size: 0.7rem;
-	width: 53%;
-	height: 80px;
-	transition: all 3s;
-	margin-top: 1%;
-	position: relative;
-	z-index: 17;
-	&__content {
-		display: flex;
-		height: 75%;
-		align-items: center;
-		margin: 0;
-	}
-	&__buttons-container {
-		width: 30%;
-		position: relative;
-		bottom: 0%;
-		display: flex;
-		justify-content: flex-start;
-	}
-	&__button {
-		height: 17px;
-		width: auto;
-		padding: 0 5%;
-		border-radius: 3px;
-		&:hover {
-			cursor: pointer;
-		}
-	}
-	&__first-button {
-		&:hover {
-			content: url('../../assets/Icons/prompt-icons/first-light-gray.png');
-		}
-	}
-	&__previous-button {
-		&:hover {
-			content: url('../../assets/Icons/prompt-icons/previous-light-gray.png');
-		}
-	}
-	&__shuffle-button {
-		&:hover {
-			content: url('../../assets/Icons/prompt-icons/shuffle-light-gray.png');
-		}
-	}
-	&__next-button {
-		&:hover {
-			content: url('../../assets/Icons/prompt-icons/next-light-gray.png');
-		}
-	}
-	&__last-button {
-		&:hover {
-			content: url('../../assets/Icons/prompt-icons/last-light-gray.png');
-		}
-	}
-	@include laptop-small() {
-		font-size: 0.7rem;
-	}
-	@include tablet-only() {
-		font-size: 0.6rem;
-	}
-	@include mobile-only() {
-		font-size: 0.5rem;
-		margin-top: 8%;
-		width: 60%;
-	}
-	@include mobile-small-only() {
-		position: relative;
-		z-index: 29;
-		font-size: 0.5rem;
-		margin-top: 15%;
-		padding-bottom: 2%;
-		top: 8%;
-		&__button {
-			font-size: 0.35rem;
-			padding: 1% 3%;
-		}
-		&__buttons-container {
-			justify-content: center;
-			margin-top: 2%;
-			width: 100%;
-		}
-	}
-}
-.red {
-	opacity: 0;
-}
-.tooltip {
-	position: relative;
-	display: inline-block;
-}
-
-.tooltip .tooltiptext {
-	visibility: hidden;
-	background-color: black;
-	border: 1px solid grey;
-	color: silver;
-	text-align: center;
-	border-radius: 6px;
-	padding: 3px;
-	position: absolute;
-	z-index: 15;
-	bottom: 120%;
-}
-
-.tooltip:hover .tooltiptext {
-	visibility: visible;
-}
